change training view models

+++ SportFinder/app/src/main/java/ru/moevm/sportfinder/screen/training/TrainingListItemVO.kt
@@ -4,8 +4,8 @@ import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 
 data class TrainingListItemVO(
-    val trainingId: Long,
+    val trainingId: Int,
     val name: String,
     val tags: ImmutableList<String> = persistentListOf(),
-    val temperature: Float? = null,
+    val temperature: Int? = null,
 )
